backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic-settings>=2.3.0
psutil>=5.9.0
pynvml>=11.5.0
sentence-transformers>=3.0.0
python-multipart>=0.0.9
